refactor: availableDateTime Status 변경 로직 수정
