+++ backend/src/main/java/com/woowacourse/ternoko/config/DatabaseInitializer.java
@@ -1,8 +1,8 @@
 package com.woowacourse.ternoko.config;
 
 import com.woowacourse.ternoko.availabledatetime.domain.AvailableDateTime;
-import com.woowacourse.ternoko.availabledatetime.domain.AvailableDateTimeRepository;
 import com.woowacourse.ternoko.availabledatetime.domain.AvailableDateTimeStatus;
+import com.woowacourse.ternoko.availabledatetime.repository.AvailableDateTimeRepository;
 import com.woowacourse.ternoko.domain.member.Coach;
 import com.woowacourse.ternoko.repository.CoachRepository;
 import com.woowacourse.ternoko.repository.CrewRepository;
